chore(apps): break out common interrupts
