+++ apps/io-node/meson.build
@@ -1,9 +1,9 @@
 io_node_inc = include_directories('include')
 io_node_src = files(
   'src' / 'app.c',
+  'src' / 'interrupts.c',
   'src' / 'main.c',
   'src' / 'stm32f3xx_hal_msp.c',
-  'src' / 'stm32f3xx_it.c',
 )
 
 io_node_elf = executable(
