--- apps/radio-node/src/stm32f3xx_it.c
@@ -1,121 +1,0 @@
-
-/**
- ******************************************************************************
- * @file    stm32f3xx_it.c
- * @brief   Interrupt Service Routines.
- ******************************************************************************
- * @attention
- *
- * Copyright (c) 2023 STMicroelectronics.
- * All rights reserved.
- *
- * This software is licensed under terms that can be found in the LICENSE file
- * in the root directory of this software component.
- * If no LICENSE file comes with this software, it is provided AS-IS.
- *
- ******************************************************************************
- */
-
-/* Includes ------------------------------------------------------------------*/
-#include "stm32f3xx_it.h"
-
-#include "FreeRTOS.h"
-#include "main.h"
-#include "task.h"
-/* Private includes ----------------------------------------------------------*/
-
-/* Private typedef -----------------------------------------------------------*/
-
-/* Private define ------------------------------------------------------------*/
-
-/* Private macro -------------------------------------------------------------*/
-
-/* Private variables ---------------------------------------------------------*/
-
-/* Private function prototypes -----------------------------------------------*/
-
-/* Private user code ---------------------------------------------------------*/
-
-/* External variables --------------------------------------------------------*/
-extern CAN_HandleTypeDef hcan;
-extern UART_HandleTypeDef huart1;
-
-/******************************************************************************/
-/*           Cortex-M4 Processor Interruption and Exception Handlers          */
-/******************************************************************************/
-/**
- * @brief This function handles Non maskable interrupt.
- */
-void NMI_Handler(void) {
-  while (1) {
-  }
-}
-
-/**
- * @brief This function handles Hard fault interrupt.
- */
-void HardFault_Handler(void) {
-  while (1) {
-  }
-}
-
-/**
- * @brief This function handles Memory management fault.
- */
-void MemManage_Handler(void) {
-  while (1) {
-  }
-}
-
-/**
- * @brief This function handles Pre-fetch fault, memory access fault.
- */
-void BusFault_Handler(void) {
-  while (1) {
-  }
-}
-
-/**
- * @brief This function handles Undefined instruction or illegal state.
- */
-void UsageFault_Handler(void) {
-  while (1) {
-  }
-}
-
-/**
- * @brief This function handles Debug monitor.
- */
-void DebugMon_Handler(void) {}
-
-/**
- * @brief This function handles System tick timer.
- */
-void SysTick_Handler(void) {
-  HAL_IncTick();
-#if (INCLUDE_xTaskGetSchedulerState == 1)
-  if (xTaskGetSchedulerState() != taskSCHEDULER_NOT_STARTED) {
-#endif /* INCLUDE_xTaskGetSchedulerState */
-    xPortSysTickHandler();
-#if (INCLUDE_xTaskGetSchedulerState == 1)
-  }
-#endif /* INCLUDE_xTaskGetSchedulerState */
-}
-
-/******************************************************************************/
-/* STM32F3xx Peripheral Interrupt Handlers                                    */
-/* Add here the Interrupt Handlers for the used peripherals.                  */
-/* For the available peripheral interrupt handler names,                      */
-/* please refer to the startup file (startup_stm32f3xx.s).                    */
-/******************************************************************************/
-
-/**
- * @brief This function handles USB high priority or CAN_TX interrupts.
- */
-void USB_HP_CAN_TX_IRQHandler(void) { HAL_CAN_IRQHandler(&hcan); }
-
-/**
- * @brief This function handles USART1 global interrupt / USART1 wake-up
- * interrupt through EXTI line 25.
- */
-void USART1_IRQHandler(void) { HAL_UART_IRQHandler(&huart1); }
