+++ common/meson.build
@@ -19,6 +19,7 @@ common_src = [
 
 common_tidy_files = [
   can_kingdom_tidy_files,
+  stm32_common_tidy_files,
   stm32_postmaster_tidy_files,
   stm32_utils_tidy_files,
 ]
