+++ common/stm32/meson.build
@@ -8,6 +8,12 @@ stm32_linker_script = configure_file(
 
 stm32_common_src = files(
   'src' / 'assert.c',
+  'src' / 'common-interrupts.c',
   'src' / 'startup_stm32f302xe.s',
   'src' / 'system_stm32f3xx.c',
+)
+
+stm32_common_tidy_files = files(
+  'src' / 'assert.c',
+  'src' / 'common-interrupts.c',
 )
